render-type: move FillPropertiesTestUtils to common utils for tests
