+++ core/src/androidTest/java/org/hisp/dhis/android/core/dataset/DataSetModelShould.java
@@ -48,20 +48,20 @@ import java.util.List;
 
 import static com.google.common.truth.Truth.assertThat;
 import static org.hisp.dhis.android.core.AndroidTestUtils.toInteger;
-import static org.hisp.dhis.android.core.utils.FillPropertiesTestUtils.CODE;
-import static org.hisp.dhis.android.core.utils.FillPropertiesTestUtils.COLOR;
-import static org.hisp.dhis.android.core.utils.FillPropertiesTestUtils.CREATED;
-import static org.hisp.dhis.android.core.utils.FillPropertiesTestUtils.DELETED;
-import static org.hisp.dhis.android.core.utils.FillPropertiesTestUtils.DESCRIPTION;
-import static org.hisp.dhis.android.core.utils.FillPropertiesTestUtils.DISPLAY_DESCRIPTION;
-import static org.hisp.dhis.android.core.utils.FillPropertiesTestUtils.DISPLAY_NAME;
-import static org.hisp.dhis.android.core.utils.FillPropertiesTestUtils.DISPLAY_SHORT_NAME;
-import static org.hisp.dhis.android.core.utils.FillPropertiesTestUtils.ICON;
-import static org.hisp.dhis.android.core.utils.FillPropertiesTestUtils.LAST_UPDATED;
-import static org.hisp.dhis.android.core.utils.FillPropertiesTestUtils.NAME;
-import static org.hisp.dhis.android.core.utils.FillPropertiesTestUtils.SHORT_NAME;
-import static org.hisp.dhis.android.core.utils.FillPropertiesTestUtils.UID;
-import static org.hisp.dhis.android.core.utils.FillPropertiesTestUtils.fillNameableModelProperties;
+import static org.hisp.dhis.android.core.data.utils.FillPropertiesTestUtils.CODE;
+import static org.hisp.dhis.android.core.data.utils.FillPropertiesTestUtils.COLOR;
+import static org.hisp.dhis.android.core.data.utils.FillPropertiesTestUtils.CREATED;
+import static org.hisp.dhis.android.core.data.utils.FillPropertiesTestUtils.DELETED;
+import static org.hisp.dhis.android.core.data.utils.FillPropertiesTestUtils.DESCRIPTION;
+import static org.hisp.dhis.android.core.data.utils.FillPropertiesTestUtils.DISPLAY_DESCRIPTION;
+import static org.hisp.dhis.android.core.data.utils.FillPropertiesTestUtils.DISPLAY_NAME;
+import static org.hisp.dhis.android.core.data.utils.FillPropertiesTestUtils.DISPLAY_SHORT_NAME;
+import static org.hisp.dhis.android.core.data.utils.FillPropertiesTestUtils.ICON;
+import static org.hisp.dhis.android.core.data.utils.FillPropertiesTestUtils.LAST_UPDATED;
+import static org.hisp.dhis.android.core.data.utils.FillPropertiesTestUtils.NAME;
+import static org.hisp.dhis.android.core.data.utils.FillPropertiesTestUtils.SHORT_NAME;
+import static org.hisp.dhis.android.core.data.utils.FillPropertiesTestUtils.UID;
+import static org.hisp.dhis.android.core.data.utils.FillPropertiesTestUtils.fillNameableModelProperties;
 
 @RunWith(AndroidJUnit4.class)
 public class DataSetModelShould extends NameableModelAbstractShould<DataSetModel, DataSet> {
